+++ lib/upgrader.dart
@@ -1,8 +1,6 @@
 /*
- * Copyright (c) 2018-2024 Larry Aasen. All rights reserved.
+ * Copyright (c) 2018-2025 Larry Aasen. All rights reserved.
  */
-
-library upgrader;
 
 export 'src/alert_style_widget.dart';
 export 'src/appcast.dart';
@@ -10,7 +8,6 @@ export 'src/itunes_search_api.dart';
 export 'src/play_store_search_api.dart';
 export 'src/upgrade_alert.dart';
 export 'src/upgrade_card.dart';
-export 'src/upgrade_device.dart';
 export 'src/upgrade_messages.dart';
 export 'src/upgrade_os.dart';
 export 'src/upgrade_state.dart';